visually od/requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comtypes==1.4.8
contourpy==1.3.1
cycler==0.12.1
filelock==3.16.1
fonttools==4.55.1
fsspec==2024.10.0
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.3
mpmath==1.3.0
networkx==3.4.2
nltk==3.9.1
numpy==2.1.3
opencv-python==4.10.0.84
packaging==24.2
pandas==2.2.3
pillow==11.0.0
psutil==6.1.0
py-cpuinfo==9.0.0
pyparsing==3.2.0
pypiwin32==223
python-dateutil==2.9.0.post0
pyttsx3==2.98
pytz==2024.2
pywin32==308
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
scipy==1.14.1
seaborn==0.13.2
setuptools==75.6.0
six==1.16.0
sympy==1.13.1
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.40
ultralytics-thop==2.0.12
urllib3==2.2.3
